@@ -2,46 +2,41 @@ source 'https://rubygems.org'
 
 # The core framework
 # https://github.com/rails/rails
-gem 'rails', '5.0.7.2'
+gem 'rails', '5.1.7'
 
-gem 'actionpack-action_caching', '~>1.2.1'
-gem 'actionpack-page_caching', '~>1.1.1'
-gem 'actionpack-xml_parser', '~>1.0.1'
-gem 'activerecord-session_store', '~>1.1.0'
-gem 'rails-observers', '~>0.1.2'
+# The backing band
+gem 'rack', '~> 2.2.3'
+gem 'rake', '~> 13.0.1'
+
+# DB adapters
+gem 'activerecord-jdbc-adapter', '~> 51.7'
+gem 'activerecord-jdbcpostgresql-adapter', '~> 51.7'
+# Oracle adapter requires ojdb7.jar; exclude from test environments.
+gem 'activerecord-oracle_enhanced-adapter', '~> 1.8.2', group: [:development, :production]
+
 gem 'responders', '~> 2.0'
 
-gem 'rake', '~> 12.3.3'
-
-gem 'rack', '~> 2.2.3'
-
-gem 'activerecord-jdbc-adapter', '~> 50.7'
-
-# Postgresql adapter
-gem 'activerecord-jdbcpostgresql-adapter', '~> 50.7'
-
-# A JSON implementation as a Ruby extension in C
 # http://flori.github.com/json/
-gem 'json', '~> 1.8.0'
+gem 'json', '~> 2.3.1-java'
 
 # CAS Strategy for OmniAuth. ETS maintains its own fork with SAML Ticket Validator capability,
 # provided by Steven Hansen.
 gem 'omniauth-cas', '~> 1.1.0', git: 'https://github.com/ets-berkeley-edu/omniauth-cas.git'
 
 # LDAP
-gem 'net-ldap', '~> 0.16.0'
+gem 'net-ldap', '~> 0.16.3'
 
 # secure_headers provides x-frame, csp and other http headers
-gem 'secure_headers', '~> 5.0.5'
+gem 'secure_headers', '~> 5.1.0'
 
 gem 'faraday', '~> 0.9.0'
 gem 'faraday_middleware', '~> 0.9.1'
-gem 'httparty', '~> 0.13.3'
+gem 'httparty', '~> 0.18.1'
 
-# OAuth2 support
-gem 'signet', '~> 0.10.0'
 gem 'google-api-client', '0.32.1'
 gem 'google_drive', '~> 3.0.5'
+# OAuth2 for Google
+gem 'signet', '~> 0.10.0'
 
 # LTI support
 gem 'ims-lti', git: 'https://github.com/instructure/ims-lti.git', branch: '1.2.x'
