@@ -24,17 +24,17 @@ describe CanvasProjectProvisionController do
 
     it_should_behave_like 'an api endpoint' do
       before { allow_any_instance_of(CanvasLti::ProjectProvision).to receive(:create_project).and_raise(RuntimeError, 'Something went wrong') }
-      let(:make_request) { post :create_project_site, :name => project_site_name }
+      let(:make_request) { post :create_project_site, params: {name: project_site_name} }
     end
 
     it_should_behave_like 'a user authenticated api endpoint' do
-      let(:make_request) { post :create_project_site, :name => project_site_name }
+      let(:make_request) { post :create_project_site, params: {name: project_site_name} }
     end
 
     context 'when user is not authorized' do
       before { allow_any_instance_of(AuthenticationStatePolicy).to receive(:can_create_canvas_project_site?).and_return(false) }
       it 'should respond with empty http 403' do
-        post :create_project_site, :name => project_site_name
+        post :create_project_site, params: {name: project_site_name}
         expect(response.status).to eq 403
         expect(response.body).to eq ''
       end
@@ -43,7 +43,7 @@ describe CanvasProjectProvisionController do
     context 'when project site name exceeds 255 characters' do
       let(:project_site_name) { (0...256).map { (65 + rand(26)).chr }.join }
       it 'should respond with bad request error' do
-        post :create_project_site, :name => project_site_name
+        post :create_project_site, params: {name: project_site_name}
         expect(response.status).to eq 400
         json_response = JSON.parse(response.body)
         expect(json_response['error']).to eq 'Project site name must be no more than 255 characters in length'
@@ -51,7 +51,7 @@ describe CanvasProjectProvisionController do
     end
 
     it 'should respond with course details' do
-      post :create_project_site, :name => project_site_name
+      post :create_project_site, params: {name: project_site_name}
       assert_response 200
       result = JSON.parse(response.body)
       expect(result).to be_an_instance_of Hash
