@@ -13,14 +13,14 @@ describe CacheController do
 
     it 'should not allow non-admin users to clear cache' do
       Rails.cache.should_not_receive(:clear)
-      get :clear, {:format => 'json'}
+      get :clear, format: 'json'
       expect(response.status).to eq(403)
       expect(response.body.blank?).to be_truthy
     end
 
     it 'should not allow non-admin users to delete a specific key' do
       expect(Rails.cache).to receive(:delete).never
-      get :delete, {key: 'Canvas::ExternalTools'}
+      get :delete, params: {key: 'Canvas::ExternalTools'}
       expect(response.status).to eq(403)
     end
   end
@@ -33,7 +33,7 @@ describe CacheController do
 
     it 'should allow superusers to clear the cache' do
       Rails.cache.should_receive(:clear).once
-      get :clear, {:format => 'json'}
+      get :clear, format: 'json'
       expect(response.status).to eq(200)
       expect(response.body).to be
       expect(response.body['cache_cleared']).to be_truthy
@@ -41,7 +41,7 @@ describe CacheController do
 
     it 'should delete a specific key' do
       expect(Rails.cache).to receive(:delete).with('Canvas::ExternalTools').and_return(true)
-      get :delete, {key: 'Canvas::ExternalTools', format: 'json'}
+      get :delete, format: 'json', params: {key: 'Canvas::ExternalTools'}
       expect(response.status).to eq(200)
       expect(response.body['deleted']).to be_truthy
     end
