@@ -78,7 +78,7 @@ describe CanvasLtiController do
       allow_any_instance_of(CanvasLti::Lti).to receive(:validate_tool_provider).and_return(lti)
     end
     it 'switches to the new authentication' do
-      post :embedded, url: 'something'
+      post :embedded, params: {url: 'something'}
       assert_response :success
       expect(session['user_id']).to eq lti_values['canvas_user_login_id']
       expect(session['canvas_user_id']).to eq lti_values['canvas_user_id']
